Rename project in default.project.json during upload
