+++ src/commands/publish.rs
@@ -1,7 +1,6 @@
 use std::path::PathBuf;
 
-use anyhow::{format_err, Context};
-use fs_err::File;
+use anyhow::Context;
 use structopt::StructOpt;
 
 use crate::{
@@ -28,28 +27,6 @@ impl PublishSubcommand {
         };
         let api = package_index.config()?.api;
         let contents = PackageContents::pack_from_path(&self.project_path)?;
-
-        let project_json_path = self.project_path.join("default.project.json");
-
-        if project_json_path.exists() {
-            let file = File::open(project_json_path)?;
-            let project_json: serde_json::Value = serde_json::from_reader(file)?;
-            let project_name = project_json
-                .get("name")
-                .and_then(|name| name.as_str())
-                .expect("Couldn't parse name in default.project.json");
-            let package_name = manifest.package.name.name();
-
-            if project_name != package_name {
-                return Err(format_err!(
-                    "The project and package names are mismatched! \
-                    The project name '{}' in `default.project.json` \
-                    must match the package name '{}' in `wally.toml`",
-                    project_name,
-                    package_name
-                ));
-            }
-        }
 
         let auth = auth_store
             .token
